+++ src/main/java/com/example/Muebleria/Dto/DetalleCompraDto.java
@@ -8,15 +8,15 @@ import lombok.Data;
 @Data
 public class DetalleCompraDto {
 
-    private Integer id;
-    private Integer compra;
+    private Long id;
+    private Long compra;
     private String producto;
     private Integer cantidad;
     private Double precioCompra;
     private Double subtotal;
     private boolean estado;
 
-    public DetalleCompraDto(Integer id, Integer compra, String producto, Integer cantidad, Double precioCompra, Double subtotal, boolean estado) {
+    public DetalleCompraDto(Long id, Long compra, String producto, Integer cantidad, Double precioCompra, Double subtotal, boolean estado) {
         this.id = id;
         this.compra = compra;
         this.producto = producto;
